+++ src/application/structs/Options.h
@@ -20,16 +20,17 @@ enum StartingArea
 
 struct Options
 {
+    std::string playerName = "Player";
     int emblemGoal = -1;
     StartingArea startingArea = NoStatingArea;
     bool lifeSanity = false;
 
-    bool sonicLifeSanity = false;
-    bool tailsLifeSanity = false;
-    bool knucklesLifeSanity = false;
-    bool amyLifeSanity = false;
-    bool bigLifeSanity = false;
-    bool gammaLifeSanity = false;
+    bool sonicLifeSanity = true;
+    bool tailsLifeSanity = true;
+    bool knucklesLifeSanity = true;
+    bool amyLifeSanity = true;
+    bool bigLifeSanity = true;
+    bool gammaLifeSanity = true;
 
     bool pinballCapsules = false;
     bool deathLinkActive = false;