+++ src/output/locationRepository/LocationRepository.cpp
@@ -14,8 +14,7 @@ LocationRepository::LocationRepository()
         {300, LocationData::UpgradeLocation(FLAG_KNUCKLES_MR_SHOVELCLAW, "Shovel Claw Upgrade (Knuckles)", 30)},
         {301, LocationData::UpgradeLocation(FLAG_KNUCKLES_MR_FIGHTINGGROVE, "Fighting Gloves Upgrade (Knuckles)", 31)},
 
-        //We grant amy the warrior's feather upgrade when getting the long hammer as well
-        {400, LocationData::UpgradeLocation(FLAG_AMY_EC_LONGHAMMER, "Warrior Feather Upgrade (Amy)", 41)},
+        {400, LocationData::UpgradeLocation(FLAG_AMY_MR_FIGHTERSFEATHER, "Warrior Feather Upgrade (Amy)", 40)},
         {401, LocationData::UpgradeLocation(FLAG_AMY_EC_LONGHAMMER, "Long Hammer Upgrade (Amy)", 41)},
 
         {500, LocationData::UpgradeLocation(FLAG_E102_EC_BOOSTER, "Jet Booster Upgrade (Gamma)", 50)},